refactor: sync bilifavlist
