+++ src/utils/Bilibili/bilifavOperate.ts
@@ -1,6 +1,7 @@
 import { getCookie, sendBVFavorite } from '../BiliOperate';
 import { getFavListBVID } from '../mediafetch/bilifavlist';
 import { humanishApiLimiter } from '../mediafetch/throttle';
+import { getBiliUser } from '../PersonalCloudAuth';
 
 const BILI_GETFAVLIST_API =
   'https://api.bilibili.com/x/v3/fav/folder/created/list-all?up_mid={mid}';
@@ -92,5 +93,24 @@ export const addToBiliFavlist = async (
           ),
         ),
       ),
+  );
+};
+
+export const syncFavlist = async (favlist: NoxMedia.Playlist) => {
+  const favid = await getOrInsertBiliFavlist(
+    (
+      await getBiliUser()
+    ).mid,
+    favlist.info.title.slice(0, 19),
+  );
+  const uniqBVIDs = Array.from(
+    favlist.songList.reduce(
+      (accumulator, currentValue) => accumulator.add(currentValue.bvid),
+      new Set() as Set<string>,
+    ),
+  );
+  await addToBiliFavlist(
+    favid,
+    uniqBVIDs.filter((val) => val.startsWith('BV')),
   );
 };
